feat: update default indexer data fetching strategy
In the readme at the time of 41e02456dc8e400f1583b334345fcbe052abbc7b commit,
it was recommended to use `ZDKFetchStrategy` instead of the deprecated
`ZoraV2Indexer` after v1 release of the hook.

Found out this when I need to explicitly override the fetching strategy
with ZDK in nft-component in order to for the result to show up. Updating
the default strategy with the recommended indexer.